requirements.txt
# Requirements files for repos.io
# For each packages not on pypi (or without the needed version), i use my own fork, to avoid lost of original repository (it happens !)

django>=1.3.1,<=1.3.99

####################
### django tools ###
####################

django-model-utils
django-redis-cache
johnny-cache
django-redisession

#django-extensions # too old, without static files from django 1.3
-e git+git://github.com/twidi/django-extensions.git#egg=django-extensions

#django-globals # my pull-request was integrated but pypi not up to date
-e git+git://github.com/twidi/django-globals.git#egg=django-globals

#django-pure-pagination ( bug if % in querystring : https://github.com/jamespacileo/django-pure-pagination/pull/6)
-e git+git://github.com/twidi/django-pure-pagination.git#egg=django-pure-pagination

# not on pypi
-e git+git://github.com/twidi/django-template-preprocessor.git#egg=django-template-preprocessor
-e git+git://github.com/twidi/django-metasettings.git#egg=django-metasettings
-e git+git://github.com/twidi/django-browsecap.git#egg=django-browsecap

####################
### python tools ###
####################
debug
textile
markdown
docutils
pysolr
beautifulsoup

# for django-template-preprocessor
termcolor

################
### backends ###
################

#github2 : commit 234e9066b5 created by ask after my issue #62 is not in 0.5.2
-e git+git://github.com/twidi/python-github2.git#egg=python-github2

###################
### django apps ###
###################

django-social-auth>=0.6,<=0.6.99
django-markup-mixin
django-notes
django-taggit
django-haystack

#saved_searches (we need more recent version than the one on pypi)
-e git+git://github.com/twidi/saved_searches.git#egg=saved_searches

############################
### background and redis ###
############################

hiredis
redis
-e git+git://github.com/twidi/redisco.git#egg=redisco

